Implementación de dimensions controller

+++ src/main/java/com/nautik/api/service/location/ZoneService.java
@@ -3,6 +3,7 @@ package com.nautik.api.service.location;
 import com.nautik.api.domain.Port;
 import com.nautik.api.domain.Zone;
 import com.nautik.api.domain.exceptions.ResourceNotFoundException;
+import com.nautik.api.domain.exceptions.ZoneConstraintViolationException;
 import com.nautik.api.domain.moorings.MooringCategory;
 import com.nautik.api.dto.location.ZoneDto;
 import com.nautik.api.dto.location.create.CreateZoneDto;
@@ -10,8 +11,10 @@ import com.nautik.api.repository.location.ZoneRepository;
 import com.nautik.api.repository.moorings.MooringCategoryRepository;
 import com.nautik.api.repository.port.PortRepository;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.exception.ConstraintViolationException;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,7 +27,7 @@ public class ZoneService {
     private final PortRepository portRepository;
     private final MooringCategoryRepository mooringCategoryRepository;
 
-    public List<ZoneDto> findByPort(Long portId){
+    public List<ZoneDto> findByPort(Long portId) {
 
         return zoneRepository.findAllByPort_Id(Math.toIntExact(portId))
                 .stream()
@@ -32,13 +35,13 @@ public class ZoneService {
                 .toList();
     }
 
-    public ZoneDto findById(Integer zoneId){
+    public ZoneDto findById(Integer zoneId) {
 
-        return modelMapper.map(zoneRepository.findZoneById(zoneId).orElseThrow(()->new ResourceNotFoundException("Zone not found")), ZoneDto.class);
+        return modelMapper.map(zoneRepository.findZoneById(zoneId).orElseThrow(() -> new ResourceNotFoundException("Zone not found")), ZoneDto.class);
     }
 
-    public ZoneDto create(Integer portId, CreateZoneDto zone){
-        Port port = portRepository.findById(portId).orElseThrow(()->new ResourceNotFoundException("Port not found"));
+    public ZoneDto create(Integer portId, CreateZoneDto zone) {
+        Port port = portRepository.findById(portId).orElseThrow(() -> new ResourceNotFoundException("Port not found"));
         Zone addZone = new Zone();
         addZone.setPort(port);
         addZone.setName(zone.getName());
@@ -46,8 +49,8 @@ public class ZoneService {
         return modelMapper.map(zoneRepository.save(addZone), ZoneDto.class);
     }
 
-    public ZoneDto update( Integer zoneId, CreateZoneDto zone ){
-        Zone searchedZone = zoneRepository.findZoneById(zoneId).orElseThrow(()->new ResourceNotFoundException("Zone not found"));
+    public ZoneDto update(Integer zoneId, CreateZoneDto zone) {
+        Zone searchedZone = zoneRepository.findZoneById(zoneId).orElseThrow(() -> new ResourceNotFoundException("Zone not found"));
 
         Zone zoneProvided = modelMapper.map(zone, Zone.class);
         zoneProvided.setId(searchedZone.getId());
@@ -55,8 +58,12 @@ public class ZoneService {
         return modelMapper.map(zoneRepository.save(zoneProvided), ZoneDto.class);
     }
 
-    public void delete(Integer zoneId){
-        Zone zoneDelete = zoneRepository.findZoneById(zoneId).orElseThrow(()->new ResourceNotFoundException("Zone not found"));
+    public void delete(Integer zoneId) {
+        Zone zoneDelete = zoneRepository.findZoneById(zoneId).orElseThrow(() -> new ResourceNotFoundException("Zone not found"));
+        if (!zoneDelete.getMooringCategories().isEmpty()) {
+            throw new ZoneConstraintViolationException();
+        }
+
         zoneRepository.delete(zoneDelete);
 
     }
